Moved all style-converter code into several classes
