+++ lib/assets/core/javascripts/cartodb3/editor/style/style-converter.js
@@ -5,6 +5,7 @@ var StyleUtils = require('./style-generation/style-generation-utils');
 var StyleConversion = require('./style-generation/style-generation-conversion');
 var StyleAggregation = require('./style-generation/style-generation-aggregation');
 var StyleGenerationSQL = require('./style-generation/style-generation-sql');
+var StyleGenerationFactory = require('./style-generation/style-generation-factory');
 var LAYER_TYPE = 'CartoDB';
 
 var CONFIG = {
@@ -12,383 +13,15 @@ var CONFIG = {
   DEFAULT_HEATMAP_COLORS: ['blue', 'cyan', 'lightgreen', 'yellow', 'orange', 'red']
 };
 
-// utilities
-function _null () { return ''; }
-
-function makeCartoCSS (obj, prefix) {
-  var css = '';
-  prefix = prefix || '';
-  for (var k in obj) {
-    css += prefix + k + ': ' + obj[k] + ';\n';
-  }
-  return css;
-}
-
-function makeColorRamp (props, isTorqueCategory) {
-  var attribute = isTorqueCategory ? 'value' : props.attribute;
-  var c = ['ramp([' + attribute + ']'];
-
-  if (props.range) {
-    if (_.isArray(props.range)) {
-      c.push('(' + props.range.join(', ') + ')');
-    } else {
-      // colorramp name
-      c.push(props.range);
-      if (props.bins) {
-        c.push(props.bins);
-      }
-    }
-  }
-  if (props.domain) {
-    if (isTorqueCategory) {
-      c.push('(' + _.map(props.domain, function (val, i) {
-        return i + 1;
-      }).join(', ') + ')');
-    } else if (props.static) {
-      // It comes from an autostyle, so we have to set the categories explicitly
-      var parsedDomain = _.filter(props.domain, function (name) {
-        return name !== '"Other"';
-      });
-      c.push('(' + parsedDomain.join(', ') + ')');
-    } else {
-      c.push('(' +
-        _.filter(
-          _.map(props.domain, function (val, i) {
-            // Maps api converts null or empty value in emtpy string
-            // and in the fill component we label them with a locale
-            // so we use the same locale to generate the cartocss
-            return val === '' ? _t('form-components.editors.fill.input-categories.null') : val;
-          }),
-          function (val) {
-            return !_.isUndefined(val);
-          }).join(', ') +
-        ')'
-      );
-      c.push('"="');
-    }
-  }
-  if (props.quantification) {
-    c.push(props.quantification.toLowerCase());
-  }
-
-  if (isTorqueCategory) {
-    c.push('"="');
-  }
-
-  return c.join(', ') + ')';
-}
-
-function makeWidthRamp (props) {
-  var c = ['ramp([' + props.attribute + ']'];
-
-  if (props.range) {
-    var min = props.range[0];
-    var max = props.range[1];
-
-    c.push('range(' + min + ', ' + max + ')');
-  }
-
-  if (props.quantification) {
-    var quantification = props.quantification.toLowerCase();
-
-    if (props.bins) {
-      c.push(quantification + '(' + props.bins + ')');
-    } else {
-      c.push(quantification);
-    }
-  }
-
-  return c.join(', ') + ')';
-}
-
-// size
-function pointSize (props) {
-  var css = {};
-  if (props.fixed !== undefined) {
-    css['marker-width'] = props.fixed;
-  } else if (props.attribute) {
-    css['marker-width'] = makeWidthRamp(props);
-  } else {
-    // throw new Error('size should contain a fixed value or an attribute')
-  }
-  return css;
-}
-
-function blending (b, geometryType, animatedType) {
-  var css = {};
-  var property = geometryType + '-comp-op';
-  if (b !== 'none' && b !== undefined && animatedType !== 'heatmap') {
-    if (animatedType === 'simple') {
-      property = 'comp-op';
-    }
-    css[property] = b;
-  }
-  return css;
-}
-
-// fill
-function pointFill (props, animationType) {
-  var css = {};
-  var color = (props && props.color) || {};
-  var isTorqueCategory = animationType && !color.fixed;
-  var markerFillOpacity = color.opacity;
-
-  if (props.size) {
-    css = pointSize(props.size);
-  }
-  if (color) {
-    if (color.fixed !== undefined) {
-      css['marker-fill'] = color.fixed;
-    } else if (color.attribute) {
-      css['marker-fill'] = makeColorRamp(color, isTorqueCategory);
-    }
-    if (color.operation) {
-      css['marker-comp-op'] = color.operation;
-    }
-
-    css['marker-fill-opacity'] = markerFillOpacity != null ? markerFillOpacity : 1;
-
-    if (StyleUtils.hasImagesSelected(props.images) || StyleUtils.hasImagesSelected(color.images)) {
-      css['marker-file'] = markerFill(props.images || color.images, color);
-    } else if (props.image || color.image) {
-      var url = props.image;
-
-      if (color.image) {
-        url = 'url(\'' + color.image + '\')';
-      }
-
-      css['marker-file'] = url;
-    }
-  }
-  if (!animationType) {
-    css['marker-allow-overlap'] = true;
-  }
-  return css;
-}
-
-function markerFill (images, color) {
-  var getFalsyCategory = function (category) {
-    return category === 0 ? '0' : "''";
-  }
-
-  if (!_.isArray(images) || !color || !color.attribute) {
-    return;
-  }
-
-  var columnName = '[' + color.attribute + ']';
-  var filesUrls = [];
-  var categoryNames = [];
-
-  _.each(images, function (image, index) {
-    if (image !== '') {
-      var urlFormat = 'url(\'' + image + '\')';
-      filesUrls.push(urlFormat);
-      if (!_.isUndefined(color.domain[index])) {
-        var category = color.domain[index] || getFalsyCategory(color.domain[index]);
-        categoryNames.push(category);
-      }
-    }
-  });
-
-  return 'ramp(' + columnName + ', (' + filesUrls.join(', ') + '), (' + categoryNames.join(', ') + '), "="' + ')';
-}
-
-function polygonFill (props) {
-  var css = {};
-  if (props.color) {
-    if (props.color.fixed !== undefined) {
-      css['polygon-fill'] = props.color.fixed;
-    } else if (props.color.attribute) {
-      css['polygon-fill'] = makeColorRamp(props.color);
-    }
-    if (props.color.operation) {
-      css['polygon-comp-op'] = props.color.operation;
-    }
-    if (_.isNumber(props.color.opacity)) {
-      css['polygon-opacity'] = props.color.opacity;
-    }
-  }
-
-  return css;
-}
-
-// stroke
-function pointStroke (props, animationType) {
-  var css = {};
-
-  if (animationType === 'heatmap') {
-    return css;
-  }
-
-  if (props.size) {
-    css['marker-line-width'] = props.size.fixed;
-  }
-  if (props.color) {
-    if (props.color.fixed !== undefined) {
-      css['marker-line-color'] = props.color.fixed;
-    } else if (props.color.attribute) {
-      css['marker-line-color'] = makeColorRamp(props.color);
-    }
-    if (_.isNumber(props.color.opacity)) {
-      css['marker-line-opacity'] = props.color.opacity;
-    }
-  }
-  return css;
-}
-
-function polygonStroke (props) {
-  var css = {};
-  if (props.size) {
-    if (props.size.fixed !== undefined) {
-      css['line-width'] = props.size.fixed;
-    } else if (props.size.attribute) {
-      css['line-width'] = makeWidthRamp(props.size);
-    }
-  }
-  if (props.color) {
-    if (props.color.fixed) {
-      css['line-color'] = props.color.fixed;
-    } else if (props.color.attribute) {
-      css['line-color'] = makeColorRamp(props.color);
-    }
-    if (_.isNumber(props.color.opacity)) {
-      css['line-opacity'] = props.color.opacity;
-    }
-  }
-  return css;
-}
-
-function isValidAttribute (attr) {
-  return attr && attr !== '';
-}
-
-function pointAnimated (props) {
-  var css = {};
-  if (isValidAttribute(props.attribute)) {
-    css['-torque-frame-count'] = props.steps;
-    css['-torque-animation-duration'] = props.duration;
-    css['-torque-time-attribute'] = '"' + props.attribute + '"';
-    if (props.isCategory) {
-      css['-torque-aggregation-function'] = '"CDB_Math_Mode(value)"';
-    } else {
-      css['-torque-aggregation-function'] = '"count(1)"';
-    }
-    css['-torque-resolution'] = props.resolution;
-    css['-torque-data-aggregation'] = props.overlap ? 'cumulative' : 'linear';
-  }
-  return css;
-}
-
-function _labels (props) {
-  var css = {};
-  if (isValidAttribute(props.attribute)) {
-    css['text-name'] = '[' + props.attribute + ']';
-    css['text-face-name'] = "'" + props.font + "'";
-    if (props.fill) {
-      css['text-size'] = props.fill.size.fixed;
-
-      if (props.fill.color.opacity != null && props.fill.color.opacity < 1) {
-        css['text-fill'] = Utils.hexToRGBA(props.fill.color.fixed, props.fill.color.opacity);
-      } else {
-        css['text-fill'] = props.fill.color.fixed;
-      }
-    }
-    css['text-label-position-tolerance'] = 0;
-    if (props.halo) {
-      css['text-halo-radius'] = props.halo.size.fixed;
-
-      if (props.halo.color.opacity != null && props.halo.color.opacity < 1) {
-        css['text-halo-fill'] = Utils.hexToRGBA(props.halo.color.fixed, props.halo.color.opacity);
-      } else {
-        css['text-halo-fill'] = props.halo.color.fixed;
-      }
-    }
-    css['text-dy'] = props.offset === undefined ? -10 : props.offset;
-    css['text-allow-overlap'] = props.overlap === undefined ? true : props.overlap;
-    css['text-placement'] = props.placement;
-    css['text-placement-type'] = 'dummy';
-  }
-  return css;
-}
-
-function imageFilters (props) {
-  var css = {};
-  if (props.ramp) {
-    css['image-filters'] = 'colorize-alpha(' + props.ramp.join(',') + ')';
-  }
-  return css;
-}
-
-var cartocssFactory = {
-  animated: {
-    point: pointAnimated,
-    line: _null,
-    polygon: _null
-  },
-
-  trails: {
-    point: trails,
-    line: _null,
-    polygon: _null
-  },
-
-  fill: {
-    point: pointFill,
-    line: _null,
-    polygon: polygonFill
-  },
-
-  stroke: {
-    point: pointStroke,
-    line: polygonStroke,
-    polygon: polygonStroke
-  },
-
-  labels: {
-    point: _labels,
-    line: _labels,
-    polygon: _labels
-  },
-
-  imageFilters: {
-    point: imageFilters,
-    line: imageFilters,
-    polygon: imageFilters
-  },
-
-  blending: {
-    point: function (attrs, animationType) { return blending(attrs, 'marker', animationType); },
-    line: function (attrs) { return blending(attrs, 'line'); },
-    polygon: function (attrs) { return blending(attrs, 'polygon'); }
-  }
-};
-
-
-
 function renderBlock (block, geometryType, animationType) {
   var css = {};
   for (var k in block) {
-    var f = cartocssFactory[k];
+    var f = StyleGenerationFactory[k];
     if (f) {
       css = _.extend(css, f[geometryType](block[k], animationType));
     }
   }
-  return makeCartoCSS(css, '  ');
-}
-
-function trails (def) {
-  var baseWidth = def.fill.size && parseInt(def.fill.size.fixed, 10);
-  var baseOpacity = def.fill.color.opacity != null ? def.fill.color.opacity : 1;
-  if (!baseWidth) return '';
-  return '\n' +
-    _.range(1, parseInt(def.animated.trails, 10) + 1)
-      .map(function (t) {
-        return '#layer[frame-offset=' + t + '] {\n' +
-          makeCartoCSS({ 'marker-width': baseWidth + 2 * t }, '  ') +
-          makeCartoCSS({ 'marker-fill-opacity': baseOpacity / (2 * t) }, '  ') +
-          '}';
-      }
-      ).join('\n');
+  return StyleUtils.convertObjToCartoCSS(css, '  ');
 }
 
 function generateCartoCSS (style, geometryType, configModel) {
@@ -432,7 +65,7 @@ function generateCartoCSS (style, geometryType, configModel) {
 
   // Animated Map trails
   if (isAnimatable && styleDef.animated.trails && styleDef.animated.trails > 0) {
-    css += cartocssFactory.trails[geometryType](styleDef);
+    css += StyleGenerationFactory['trails'][geometryType](styleDef);
   }
 
   return css;
