+++ lib/assets/core/javascripts/cartodb3/editor/style/style-generation/style-filters-generation.js
@@ -1,0 +1,11 @@
+/**
+ *  Image-filter CartoCSS generation for heatmap layers
+ */
+
+module.exports = function (props) {
+  var css = {};
+  if (props.ramp) {
+    css['image-filters'] = 'colorize-alpha(' + props.ramp.join(',') + ')';
+  }
+  return css;
+};
