+++ lib/assets/core/javascripts/cartodb3/editor/style/style-generation/style-generation-utils.js
@@ -33,5 +33,18 @@ module.exports = {
 
   isLabelsEnabled: function (styleDef) {
     return styleDef.labels && styleDef.labels.enabled && !!styleDef.labels.enabled;
+  },
+
+  isStyleAttributeValid: function (attribute) {
+    return attribute && attribute !== '';
+  },
+
+  convertObjToCartoCSS: function (obj, prefix) {
+    var css = '';
+    prefix = prefix || '';
+    for (var k in obj) {
+      css += prefix + k + ': ' + obj[k] + ';\n';
+    }
+    return css;
   }
 };
